pyproject.toml
[project]
name = "recipe-recommendation-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.9"
dependencies = [
    "langchain>=0.3.27",
    "langchain-openai>=0.3.33",
    "python-dotenv>=1.1.1",
]
